refactor : rename react query hook
